% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch_01_00.R
\name{kmeans_hobbit}
\alias{kmeans_hobbit}
\title{General k-means function with all algo options}
\usage{
kmeans_hobbit(data, k, max_iter = 20, tol = 1e-04, method = "hw")
}
\arguments{
\item{method}{}
}
\description{
General k-means function with all algo options
}
