% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansj.R
\name{kmeansj}
\alias{kmeansj}
\title{General k-means function with all algo options}
\usage{
kmeansj(data, k, max_iter = 20, tol = 1e-04, method = "hw")
}
\arguments{
\item{data}{data.frame}

\item{k}{integer}

\item{max_iter}{integer}

\item{tol}{positive numeric}

\item{method}{"lloyd", "mcq" or "hw"}
}
\value{
list, see vignette for more information
}
\description{
General k-means function with all algo options
}
