% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{silhouette}
\alias{silhouette}
\title{Silhouette Score}
\usage{
silhouette(fit, data)
}
\arguments{
\item{fit}{return object from 'kmeansj()'}

\item{data}{data.frame}
}
\value{
silhouette score of 'fit'
}
\description{
Silhouette Score
}
