% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{km_predict}
\alias{km_predict}
\title{Cluster prediction on test data}
\usage{
km_predict(data, fit)
}
\arguments{
\item{data}{unseen test data}

\item{fit}{fitted model}
}
\description{
Cluster prediction on test data
}
